more sanity checking, and make type sizes explicit
in serve.go, where the privileged operations are done, check that the parameters make sense. that's the whole point of doing it in a separate process.

make buildID int32, that's what it is in postgres anyway.
make UIDs and GIDs uint32, that's what stat seems to return.

+++ httpserve.go
@@ -57,7 +57,7 @@ func servehttp(args []string) {
 	initDingDataDir()
 
 	// be cautious
-	if config.IsolateBuilds.Enabled && (os.Getuid() != config.IsolateBuilds.DingUID || os.Getgid() != config.IsolateBuilds.DingGID) {
+	if config.IsolateBuilds.Enabled && (uint32(os.Getuid()) != config.IsolateBuilds.DingUID || uint32(os.Getgid()) != config.IsolateBuilds.DingGID) {
 		log.Fatalln("not running under expected uid/gid")
 	}
